# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Brain Battle API",
    "version": "0.1.0"
  },
  "paths": {
    "/auth/teacher": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login Teacher",
        "operationId": "login_teacher_auth_teacher_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeacherLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/student": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login Student",
        "operationId": "login_student_auth_student_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/questions": {
      "get": {
        "tags": [
          "questions"
        ],
        "summary": "Get Questions",
        "operationId": "get_questions_questions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/QuestionResponse"
                  },
                  "type": "array",
                  "title": "Response Get Questions Questions Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "questions"
        ],
        "summary": "Create Question",
        "operationId": "create_question_questions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/questions/{question_id}": {
      "patch": {
        "tags": [
          "questions"
        ],
        "summary": "Update Question",
        "operationId": "update_question_questions__question_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Question Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "questions"
        ],
        "summary": "Delete Question",
        "operationId": "delete_question_questions__question_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Question Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/rooms": {
      "post": {
        "tags": [
          "rooms"
        ],
        "summary": "Create Room",
        "operationId": "create_room_rooms_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/rooms/{room_id}/start": {
      "patch": {
        "tags": [
          "rooms"
        ],
        "summary": "Start Room",
        "operationId": "start_room_rooms__room_id__start_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Room Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/rooms/{room_id}/teams": {
      "post": {
        "tags": [
          "teams"
        ],
        "summary": "Create Team",
        "operationId": "create_team_rooms__room_id__teams_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Room Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamJoinRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/rooms/by-pin/{pin}/teams": {
      "post": {
        "tags": [
          "teams"
        ],
        "summary": "Create Team By Room Pin",
        "operationId": "create_team_by_room_pin_rooms_by_pin__pin__teams_post",
        "deprecated": true,
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pin",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pin"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamJoinRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/teams/{team_id}/students": {
      "post": {
        "tags": [
          "teams"
        ],
        "summary": "Join Team",
        "operationId": "join_team_teams__team_id__students_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Team Id"
            }
          },
          {
            "name": "studentNumber",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Studentnumber"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/teams/{team_id}/progress": {
      "get": {
        "tags": [
          "game"
        ],
        "summary": "Get Team Progress",
        "operationId": "get_team_progress_teams__team_id__progress_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Team Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamProgressResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/teams/{team_id}/answers/batch": {
      "post": {
        "tags": [
          "game"
        ],
        "summary": "Submit Answers Batch",
        "operationId": "submit_answers_batch_teams__team_id__answers_batch_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Team Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchAnswerSubmitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchAnswerSubmitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/teams/{team_id}/answers/{question_id}": {
      "post": {
        "tags": [
          "game"
        ],
        "summary": "Submit Answer",
        "operationId": "submit_answer_teams__team_id__answers__question_id__post",
        "deprecated": true,
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Team Id"
            }
          },
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Question Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerSubmitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/rooms/{room_id}/ranking": {
      "get": {
        "tags": [
          "rooms"
        ],
        "summary": "Get Ranking",
        "operationId": "get_ranking_rooms__room_id__ranking_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Room Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RankingResponse"
                  },
                  "title": "Response Get Ranking Rooms  Room Id  Ranking Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnswerSubmitRequest": {
        "properties": {
          "submittedAnswer": {
            "type": "string",
            "title": "Submittedanswer"
          }
        },
        "type": "object",
        "required": [
          "submittedAnswer"
        ],
        "title": "AnswerSubmitRequest"
      },
      "AuthTokenResponse": {
        "properties": {
          "accessToken": {
            "type": "string",
            "title": "Accesstoken"
          },
          "tokenType": {
            "type": "string",
            "title": "Tokentype",
            "default": "bearer"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "expiresIn": {
            "type": "integer",
            "title": "Expiresin"
          },
          "studentId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Studentid"
          },
          "teamId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Teamid"
          }
        },
        "type": "object",
        "required": [
          "accessToken",
          "role",
          "expiresIn"
        ],
        "title": "AuthTokenResponse"
      },
      "BatchAnswerItem": {
        "properties": {
          "questionId": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Questionid"
          },
          "submittedAnswer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submittedanswer"
          }
        },
        "type": "object",
        "required": [
          "questionId"
        ],
        "title": "BatchAnswerItem"
      },
      "BatchAnswerSubmitRequest": {
        "properties": {
          "answers": {
            "items": {
              "$ref": "#/components/schemas/BatchAnswerItem"
            },
            "type": "array",
            "maxItems": 10,
            "title": "Answers"
          }
        },
        "type": "object",
        "title": "BatchAnswerSubmitRequest"
      },
      "BatchAnswerSubmitResponse": {
        "properties": {
          "teamId": {
            "type": "integer",
            "title": "Teamid"
          },
          "submissionRound": {
            "type": "integer",
            "title": "Submissionround"
          },
          "correctCount": {
            "type": "integer",
            "title": "Correctcount"
          },
          "totalQuestions": {
            "type": "integer",
            "title": "Totalquestions"
          },
          "finished": {
            "type": "boolean",
            "title": "Finished"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionProgressResponse"
            },
            "type": "array",
            "title": "Questions"
          },
          "rotationQuestionIds": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Rotationquestionids"
          },
          "nextQuestionId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextquestionid"
          },
          "gradedCount": {
            "type": "integer",
            "title": "Gradedcount"
          }
        },
        "type": "object",
        "required": [
          "teamId",
          "submissionRound",
          "correctCount",
          "totalQuestions",
          "finished",
          "questions",
          "rotationQuestionIds",
          "nextQuestionId",
          "gradedCount"
        ],
        "title": "BatchAnswerSubmitResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QuestionRequest": {
        "properties": {
          "questionNumber": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Questionnumber"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "maxSubmitCount": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Maxsubmitcount"
          }
        },
        "type": "object",
        "required": [
          "questionNumber",
          "answer",
          "maxSubmitCount"
        ],
        "title": "QuestionRequest"
      },
      "QuestionResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "questionNumber": {
            "type": "integer",
            "title": "Questionnumber"
          },
          "maxSubmitCount": {
            "type": "integer",
            "title": "Maxsubmitcount"
          }
        },
        "type": "object",
        "required": [
          "id",
          "questionNumber",
          "maxSubmitCount"
        ],
        "title": "QuestionResponse"
      },
...
      "RoomResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "pin": {
            "type": "string",
            "title": "Pin"
          },
          "started": {
            "type": "boolean",
            "title": "Started"
          }
        },
        "type": "object",
        "required": [
          "id",
          "pin",
          "started"
        ],
        "title": "RoomResponse"
      },
      "StudentLoginRequest": {
        "properties": {
          "teamPin": {
            "type": "string",
            "maxLength": 6,
            "minLength": 6,
            "title": "Teampin"
          },
          "studentNumber": {
            "type": "string",
            "minLength": 1,
            "title": "Studentnumber"
          }
        },
        "type": "object",
        "required": [
          "teamPin",
          "studentNumber"
        ],
        "title": "StudentLoginRequest"
      },
      "StudentResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "studentNumber": {
            "type": "string",
            "title": "Studentnumber"
          },
          "teamId": {
            "type": "integer",
            "title": "Teamid"
          }
        },
        "type": "object",
        "required": [
          "id",
          "studentNumber",
          "teamId"
        ],
        "title": "StudentResponse"
      },
      "TeacherLoginRequest": {
        "properties": {
          "accessCode": {
            "type": "string",
            "minLength": 1,
            "title": "Accesscode"
          }
        },
        "type": "object",
        "required": [
          "accessCode"
        ],
        "title": "TeacherLoginRequest"
      },
      "TeamJoinRequest": {
        "properties": {
          "teamName": {
            "type": "string",
            "title": "Teamname"
          }
        },
        "type": "object",
        "required": [
          "teamName"
        ],
        "title": "TeamJoinRequest"
      },
      "TeamProgressResponse": {
        "properties": {
          "teamId": {
            "type": "integer",
            "title": "Teamid"
          },
          "submissionRound": {
            "type": "integer",
            "title": "Submissionround"
          },
          "correctCount": {
            "type": "integer",
            "title": "Correctcount"
          },
          "totalQuestions": {
            "type": "integer",
            "title": "Totalquestions"
          },
          "finished": {
            "type": "boolean",
            "title": "Finished"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionProgressResponse"
            },
            "type": "array",
            "title": "Questions"
          },
          "rotationQuestionIds": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Rotationquestionids"
          },
          "nextQuestionId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextquestionid"
          }
        },
        "type": "object",
        "required": [
          "teamId",
          "submissionRound",
          "correctCount",
          "totalQuestions",
          "finished",
          "questions",
          "rotationQuestionIds",
          "nextQuestionId"
        ],
        "title": "TeamProgressResponse"
      },
      "TeamResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "pin": {
            "type": "string",
            "title": "Pin"
          },
          "currentCount": {
            "type": "integer",
            "title": "Currentcount"
          },
          "submissionRound": {
            "type": "integer",
            "title": "Submissionround"
          },
          "finished": {
            "type": "boolean",
            "title": "Finished"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "pin",
          "currentCount",
          "submissionRound",
          "finished"
        ],
        "title": "TeamResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}